+++ pcf8563.h
@@ -97,12 +97,12 @@ extern "C" {
 
     /* These should be provided by the HAL. */
     typedef struct {
-        int32_t(* read)(void *handle, uint8_t address, uint8_t reg, uint8_t *buffer, uint16_t size);
-        int32_t(* write)(void *handle, uint8_t address, uint8_t reg, const uint8_t *buffer, uint16_t size);
+        int8_t(* read)(void *handle, uint8_t address, uint8_t reg, uint8_t *buffer, uint16_t size);
+        int8_t(* write)(void *handle, uint8_t address, uint8_t reg, const uint8_t *buffer, uint16_t size);
         void *handle;
     } pcf8563_t;
 
-    typedef int32_t pcf8563_err_t;
+    typedef int8_t pcf8563_err_t;
 
     pcf8563_err_t pcf8563_init(const pcf8563_t *pcf);
     pcf8563_err_t pcf8563_read(const pcf8563_t *pcf, struct tm *time);
